fix, app - refactoring
